+++ CHANGES.md
@@ -7,6 +7,12 @@ This release adds support for using `--scie-pbs-stripped` in combination with
 
 * Support PBS free-threaded `install_only(_stripped)`. (#3197)
 
+## 2.96.2
+
+This release fixes `--scie-only` for foreign platforms.
+
+* Fix `--scie-only` for foreign platforms. (#3196)
+
 ## 2.96.1
 
 This release fixes `pex3 lock sync` to handle multiple input requirements for the same project